+++ changelog.md
@@ -1,6 +1,10 @@
 Codion Change Log
 =================
 
+## 0.18.34
+### is.codion.framework.domain
+- Column.Getter and Setter renamed GetValue and SetParameter respectively.
+
 ## 0.18.33
 ### is.codion.swing.common.ui
 - Icons.icon() renamed get().